test(pipeline): cover account-resolve nudge gaps
Split the oversized nudge test file into a shared fixtures module
(AccountResolveActions.fixtures.ts) and a sibling multi-tier file
(AccountResolveActions.MultiTier.test.ts) so each stays under the
600-line test cap, then add two behaviour tests that close the
account-resolve branch-coverage gap:

- a passive miss when an accounts container's only identifier is an
  unusable placeholder ('1', rejected by isUsableIdentifier for
  length < 2), so the SPA is still nudged rather than mistaking the
  skeleton row for a resolved account (Wait.ts findFirstIdInPool).
- a clean tier-3 abort when the navigation itself fails, so the
  nudge never probes for accounts traffic on a page that never
  loaded (Nudge.ts tierHrefNavigate).

Lifts gated branch coverage back over the 95.1% floor. No
production change; the nudge fix and its wiring are unchanged.
