+++ src/Tests/Unit/Pipeline/Mediator/AccountResolve/AccountResolveActions.test.ts
@@ -8,7 +8,6 @@
  * ids from each, so the new phase is PURE GENERIC — no per-bank code.
  */
 
-import type { SelectorCandidate } from '../../../../../Scrapers/Base/Config/LoginConfigTypes.js';
 import {
   ACCOUNT_RESOLVE_BUDGET_MS,
   executeAccountResolveAction,
@@ -19,18 +18,12 @@ import {
 import {
   type IElementMediator,
   type IRaceResult,
-  NOT_FOUND_RESULT,
 } from '../../../../../Scrapers/Pipeline/Mediator/Elements/ElementMediator.js';
 import type { IDiscoveredEndpoint } from '../../../../../Scrapers/Pipeline/Mediator/Network/NetworkDiscoveryTypes.js';
-import { WK_DASHBOARD } from '../../../../../Scrapers/Pipeline/Registry/WK/DashboardWK.js';
 import type { IPipelineContext } from '../../../../../Scrapers/Pipeline/Types/PipelineContext.js';
-import { isOk, type Procedure, succeed } from '../../../../../Scrapers/Pipeline/Types/Procedure.js';
+import { isOk, type Procedure } from '../../../../../Scrapers/Pipeline/Types/Procedure.js';
 import { makeMockContext } from '../../Infrastructure/MockFactories.js';
-
-/** Shared no-op nudge result — a `found:false` success matching the real
- *  `resolveAndClick` contract (`Promise<Procedure<IRaceResult>>`) without
- *  driving an id into the pool. */
-const NUDGE_NOOP_RESULT = succeed(NOT_FOUND_RESULT);
+import { makeCapture, NUDGE_NOOP_RESULT } from './AccountResolveActions.fixtures.js';
 
 /** Args bundle for the synthetic-pool mediator factory. */
 interface IPoolMediatorArgs {
@@ -101,32 +94,6 @@ function makePoolMediator(args: IPoolMediatorArgs): IElementMediator {
       return Promise.resolve(NUDGE_NOOP_RESULT);
     },
   } as unknown as IElementMediator;
-}
-
-/** Args for `makeCapture`. */
-interface IMakeCaptureArgs {
-  readonly url: string;
-  readonly method: 'GET' | 'POST';
-  readonly responseBody: unknown;
-  readonly postData?: string;
-}
-
-/**
- * Build a synthetic discovered endpoint.
- * @param args - Capture args.
- * @returns Synthetic IDiscoveredEndpoint.
- */
-function makeCapture(args: IMakeCaptureArgs): IDiscoveredEndpoint {
-  return {
-    url: args.url,
-    method: args.method,
-    postData: args.postData ?? '',
-    responseBody: args.responseBody,
-    contentType: 'application/json',
-    requestHeaders: {},
-    responseHeaders: {},
-    timestamp: 100,
-  };
 }
 
 describe('ACCOUNT_RESOLVE_BUDGET_MS', () => {
@@ -609,187 +576,35 @@ describe('executeAccountResolvePre — same-URL SPA cards-view nudge', () => {
       }
     }
   });
-});
 
-/** Args for {@link makeHiddenMenuMediator}. */
-interface IHiddenMenuArgs {
-  /** Id-bearing accounts capture revealed once the cards view loads. */
-  readonly reveal: IDiscoveredEndpoint;
-  /** Tier-2: a TRANSACTIONS click reveals the pool only after MENU_EXPAND. */
-  readonly revealOnExpand?: boolean;
-  /** Tier-3 input: hrefs returned by `collectAllHrefs`. */
-  readonly hrefs?: readonly string[];
-  /** Tier-3: `navigateTo` reveals the pool (a real txn href existed). */
-  readonly revealOnNavigate?: boolean;
-}
-
-/**
- * Build a hidden-menu SPA mediator: the transactions link resolves only
- * AFTER the collapsed menu is expanded (tier-2), or via a txn-page href
- * (tier-3). Models the live Isracard/Amex account-resolve failures the
- * single-shot nudge could not recover — a lone TRANSACTIONS click lands
- * on nothing, leaving the pool empty so POST fails loud. Clicks are
- * matched by reference identity against the exact WK groups the
- * production nudge passes.
- * @param args - Reveal capture + which recovery path is wired.
- * @returns Mediator stub plus the recorded navigateTo targets.
- */
-function makeHiddenMenuMediator(args: IHiddenMenuArgs): {
-  mediator: IElementMediator;
-  navTargets: string[];
-} {
-  const revealedPool: readonly IDiscoveredEndpoint[] = [args.reveal];
-  const menuGroup = WK_DASHBOARD.MENU_EXPAND as unknown as readonly SelectorCandidate[];
-  const okFound: Procedure<IRaceResult> = succeed({ ...NOT_FOUND_RESULT, found: true });
-  const okVoid: Procedure<void> = succeed(undefined);
-  const navTargets: string[] = [];
-  let pool: readonly IDiscoveredEndpoint[] = [];
-  let isMenuExpanded = false;
-
-  const mediator = {
-    network: {
-      /**
-       * Snapshot the current pre-nav pool.
-       * @returns Captured endpoints.
-       */
-      getPreNavCaptures: (): readonly IDiscoveredEndpoint[] => pool,
-      /**
-       * Apply the id predicate to the current pool.
-       * @param _timeoutMs - Ignored budget.
-       * @param predicate - Id-bearing capture selector.
-       * @returns Match endpoint or false.
-       */
-      waitForFirstId: (
-        _timeoutMs: number,
-        predicate: (p: readonly IDiscoveredEndpoint[]) => IDiscoveredEndpoint | false,
-      ): Promise<IDiscoveredEndpoint | false> => {
-        const match = predicate(pool);
-        return Promise.resolve(match);
+  it('treats an accounts container whose only id is an unusable placeholder as a passive miss', async () => {
+    // An accounts-shaped container (cardsList) IS recognised as an endpoint,
+    // but its sole record carries only a placeholder identifier ('1', a
+    // position-index sentinel that isUsableIdentifier rejects for length < 2).
+    // findFirstIdInPool therefore sees the endpoint yet zero usable ids, so the
+    // SPA is still nudged to fire the real GetCardList rather than mistaking the
+    // skeleton row for a resolved account. The nudge reveals idCapture; POST resolves.
+    const idlessContainer = makeCapture({
+      url: 'https://web.isracard.example/ocp/statuspage/DigitalV3.StatusPage/GetCardList',
+      method: 'POST',
+      responseBody: {
+        data: { cardsList: [{ cardSuffix: '1', heading: 'Statement', currencyText: 'ILS' }] },
       },
-      /**
-       * Resolve the first pooled capture whose URL matches a pattern.
-       * @param patterns - Accounts-API URL patterns.
-       * @returns Matching endpoint or false.
-       */
-      waitForTraffic: (patterns: readonly RegExp[]): Promise<IDiscoveredEndpoint | false> => {
-        const match = pool.find(capture => patterns.some(p => p.test(capture.url)));
-        return Promise.resolve(match ?? false);
-      },
-    },
-    /**
-     * Smart-wait mock — awaits the custom wait so test stubs run.
-     * @param cw - Caller-supplied custom wait promise.
-     * @returns True after the custom wait settles.
-     */
-    raceWithNetworkIdle: async (cw: Promise<unknown>): Promise<true> => {
-      try {
-        await cw;
-      } catch {
-        /* swallow */
-      }
-      return true as const;
-    },
-    /**
-     * Reference-identity click handler. A MENU_EXPAND click marks the menu
-     * open; a later TRANSACTIONS click then reveals the pool (tier-2).
-     * Every other click is a best-effort no-op (the collapsed-menu miss).
-     * @param candidates - The exact WK group the production nudge passed.
-     * @returns Found sentinel for handled groups, no-op result otherwise.
-     */
-    resolveAndClick: (
-      candidates: readonly SelectorCandidate[],
-    ): Promise<Procedure<IRaceResult>> => {
-      const isMenuClick = candidates === menuGroup;
-      if (isMenuClick) isMenuExpanded = true;
-      const isTxnReveal = candidates === WK_DASHBOARD.TRANSACTIONS && args.revealOnExpand === true;
-      if (isTxnReveal && isMenuExpanded) pool = revealedPool;
-      const isHandled = isMenuClick || (isTxnReveal && isMenuExpanded);
-      return Promise.resolve(isHandled ? okFound : NUDGE_NOOP_RESULT);
-    },
-    /**
-     * Expose the configured hrefs for the tier-3 href scan.
-     * @returns Hrefs present on the page.
-     */
-    collectAllHrefs: (): Promise<readonly string[]> => Promise.resolve(args.hrefs ?? []),
-    /**
-     * Report the current SPA URL used to resolve relative hrefs.
-     * @returns Current page URL.
-     */
-    getCurrentUrl: (): string => 'https://web.isracard.example/ocp',
-    /**
-     * Record a navigation target and, when wired, reveal the pool — models
-     * a real transactions-page href existing for the tier-3 navigation.
-     * @param url - Absolute URL the nudge navigated to.
-     * @returns Best-effort navigation success sentinel.
-     */
-    navigateTo: (url: string): Promise<Procedure<void>> => {
-      navTargets.push(url);
-      if (args.revealOnNavigate === true) pool = revealedPool;
-      return Promise.resolve(okVoid);
-    },
-  } as unknown as IElementMediator;
-  return { mediator, navTargets };
-}
-
-describe('executeAccountResolvePre — multi-tier hidden-menu nudge (regression firing tests)', () => {
-  const reveal = makeCapture({
-    url: 'https://web.isracard.example/ocp/statuspage/DigitalV3.StatusPage/GetCardList',
-    method: 'POST',
-    responseBody: { data: { cardsList: [{ cardSuffix: 'FAKE_C01', accountNumber: '111' }] } },
-  });
-
-  /**
-   * Wrap a hidden-menu mediator in a has:true pipeline context.
-   * @param mediator - Hidden-menu mediator stub.
-   * @returns Context with the mediator present.
-   */
-  function ctxFor(mediator: IElementMediator): IPipelineContext {
-    return { ...makeMockContext(), mediator: { has: true, value: mediator } };
-  }
-
-  it('resolves accounts when the transactions link is hidden behind a collapsed menu', async () => {
-    // RED on the single-shot nudge: a lone TRANSACTIONS click lands on
-    // nothing (menu collapsed) → GetCardList never fires → empty pool →
-    // POST fails loud. The multi-tier nudge expands the menu first.
-    const { mediator } = makeHiddenMenuMediator({ reveal, revealOnExpand: true });
-    const ctx = ctxFor(mediator);
+    });
+    const mediator = makePoolMediator({
+      captures: [idlessContainer],
+      onClickCaptures: [idCapture],
+    });
+    const ctx = ctxWith(mediator);
     const pre = await executeAccountResolvePre(ctx);
     expect(pre.success).toBe(true);
     const post = await executeAccountResolvePost(ctx);
     expect(post.success).toBe(true);
     if (isOk(post)) {
       expect(post.value.accountDiscovery.has).toBe(true);
+      if (post.value.accountDiscovery.has) {
+        expect(post.value.accountDiscovery.value.ids.length).toBeGreaterThan(0);
+      }
     }
-  });
-
-  it('resolves accounts by navigating a txn-page href when no link resolves', async () => {
-    // RED on the single-shot nudge: no clickable link resolves at all →
-    // tier-3 navigates directly to a discovered transactions-page href.
-    const { mediator, navTargets } = makeHiddenMenuMediator({
-      reveal,
-      hrefs: ['https://web.isracard.example/ocp/transactions'],
-      revealOnNavigate: true,
-    });
-    const ctx = ctxFor(mediator);
-    const pre = await executeAccountResolvePre(ctx);
-    expect(pre.success).toBe(true);
-    const post = await executeAccountResolvePost(ctx);
-    expect(post.success).toBe(true);
-    expect(navTargets).toContain('https://web.isracard.example/ocp/transactions');
-  });
-
-  it('never navigates to a same-text login-redirect href (decoy rejected)', async () => {
-    // Amex signature: the only same-text hrefs point at /personalarea/login/.
-    // The txn-pattern filter must reject them so the nudge never navigates
-    // to the login page (which would destroy the authenticated session).
-    const { mediator, navTargets } = makeHiddenMenuMediator({
-      reveal,
-      hrefs: ['https://web.isracard.example/personalarea/login/'],
-      revealOnNavigate: true,
-    });
-    const ctx = ctxFor(mediator);
-    const pre = await executeAccountResolvePre(ctx);
-    expect(pre.success).toBe(true);
-    expect(navTargets).toHaveLength(0);
   });
 });
